+++ brepair/irepair0.c
@@ -55,7 +55,7 @@ int PRNL = 0;  // print progress on text scan
 
 float factor = 0.75; // 1/extra space overhead; set closer to 1 for smaller and
          // slower execution
-int minsize = 256; // to avoid many reallocs at small sizes, should be ok as is
+long minsize = 256; // to avoid many reallocs at small sizes, should be ok as is
 
 relong u; // |text| and later current |C| with gaps
